Actualizando Transacciones - Parte 2

+++ ManejoPresupuesto/Models/TransaccionActualizacionViewModel.cs
@@ -3,6 +3,6 @@
     public class TransaccionActualizacionViewModel : TransaccionCreacionViewModel
     {
         public int CuentaAnteriorId { get; set; }
-        public int MontoAnterior { get; set; }
+        public decimal MontoAnterior { get; set; }
     }
 }
